requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
Faker==40.5.1
greenlet==3.3.2
idna==3.11
iniconfig==2.3.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.19.2
pytest==9.0.2
pytest-assume==2.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
xdist==0.0.2
